add patches/ directory, add no-haskell patch
bacontools refuse to build and install on certain systems.
Specifically, current Haskell packages for Raspbian 8 cannot
build `bananaglee`, since cabal fails with a "double free or
corruption" error, and runghc is, apparently, not suppplied for the
ARMv7l architecture. `0001-no-haskell.patch` fixes it, disabling the
build and install of `bananaglee`.